+++ koolreport/autoload.php
@@ -22,7 +22,6 @@ spl_autoload_register(function ($class_name) {
         //try to load in packages
         $class_name = str_replace("\\","/",$class_name);
         $filePath = str_replace("\\","/",dirname(dirname(__FILE__)))."/".str_replace("koolreport", "koolreport/packages", $class_name).'.php';
-        
         if(is_file($filePath))
         {
             require_once $filePath;